Some structure for Flee
I know the struct EnemyLocation will need to be moved or reworked - just
in as a temporary struct to be able to have functions act properly.

Sam and I briefly talked about passing the enemy object into update so
that we would be able to use get locations from it. That could resolve
my need for an enemyLocation struct.